multi_llm_chatbot_backend/requirements.txt
# Core FastAPI framework
fastapi
uvicorn[standard]
python-multipart

# HTTP client for LLM APIs
httpx

# Document processing
PyPDF2
docx2txt
python-docx

# Environment configuration
python-dotenv
pyyaml

# Vector database and embeddings
chromadb
sentence-transformers

# Natural language processing
nltk
tiktoken

# PDF generation and export
reportlab

# Database (MongoDB)
pymongo
motor

# Authentication and security
passlib[bcrypt]
# `bcrypt` pinned for compat.
bcrypt~=4.0
python-jose[cryptography]

# Data validation (required for EmailStr in Pydantic)
email-validator
python-docx
python-multipart
